+++ scripts/storage.js
@@ -1,4 +1,4 @@
-var changes = 43; // Update this if readingPagesConfig is updated
+var changes = 44; // Update this if readingPagesConfig is updated
 
 var readingPagesConfig = {
 	readingConfigName: '',
@@ -67,6 +67,10 @@ var storageDefault = {
 		controllerDeadZone: 0.06,
 		startInFullScreen: false,
 		renderMaxWidth: 12000,
+		checkReleases: true,
+		checkPreReleases: true,
+		lastCheckedRelease: '',
+		lastCheckedReleaseTime: 0,
 		trackingSites: {
 			wildcard: {
 				favorite: false,